StudyPES_material_retrival/requirements.txt
PyMuPDF>=1.23.0
requests>=2.31.0
google-generativeai>=0.3.0
python-pptx>=0.6.21
python-docx>=0.8.11
